parserApp/requirements.txt
Django
django-crispy-forms==1.14.0
django-environ==0.9.0
django-extensions==3.2.1
django-js-asset==2.0.0
psycopg2-binary
gunicorn
requests==2.28.1
python-dateutil==2.8.2
